+++ flowcraft/generator/components/assembly.py
@@ -216,3 +216,29 @@ class Abyss(Process):
             "version": "2.1.1",
             "scratch": "true"
         }}
+
+class Unicycler(Process):
+    """Unicycler process template interface
+
+    This process is set with:
+
+        - ``input_type``: fastq
+        - ``output_type``: assembly
+        - ``ptype``: assembly
+
+    """
+
+    def __init__(self, **kwargs):
+
+        super().__init__(**kwargs)
+
+        self.input_type = "fastq"
+        self.output_type = "fasta"
+        self.link_start.append("gfa1")
+
+        self.directives = {"unicycler": {
+            "cpus": 4,
+            "container": "quay.io/biocontainers/unicycler",
+            "version": "0.4.7--py36hdbcaa40_0",
+            "scratch": "true"
+        }}