+++ changelog.md
@@ -23,11 +23,14 @@ which is particularly useful in very large workflows.
 
 ### Components changes
 
+- Added new `disableRR` param in the `spades` component that disables repeat
+resolution.
 - Updated images for components `mash_dist`, `mash_screen` and 
 `mapping_patlas`.
 
 ### New components
 - Added component `fast_ani`.
+- Added component `abyss`.
 
 ### Minor/Other changes