+++ changelog.md
@@ -17,6 +17,8 @@ resolution
 ### New components
 
 - Added component `abyss`.
+- Added component `bandage`.
+- Added component `unicycler`.
 
 ### Minor/Other changes